+++ packages/core/assets/plugins/index.ts
@@ -19,7 +19,7 @@ export async function runAppPlugins (
       if (types.ready && __state !== "start") return
       plugin.__state = type
 
-      const { DESKTOP, MOBILE, WEB, TARGET, DEV, PROD } = this.env
+      const { DESKTOP, MOBILE, WEB, TARGET, DEV } = this.env
       const featureIsSupported = await isPluginFeatureSupported.call({ WEB, DESKTOP: DESKTOP ? TARGET : false,  MOBILE: MOBILE ? TARGET : false, DEV: !!DEV, PROD: !DEV }, plugin, type)
       if (!featureIsSupported) return