+++ packages/core/services/types.ts
@@ -2,7 +2,7 @@ export type Service = {
     name: string,
     src: string,
     description?: string,
-    host?: string | "0.0.0.0"
+    public?: boolean,
     port?: number,
     publish?: any
 }