+++ packages/core/CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Allow for popup window configuration, bundling, and runtime control in plugins.
 
 ### Changed
+- Specify `public` directly instead of complete control over the `host` for services.
 - Handle `isSupported` queries per feature on the plugin. 
 - Provide a `start`, `ready`, and `quit` callback on plugins across all targets (note: dev only for web and mobile).
 - Added `PAGES` global variable to simplify navigation on multi-page applications.